Report missing rules during deletion
Although the implementation for AuditClient.Delete is only used to back
AuditClient.DeleteAll, we'd like to be able to delete individual rules.
This commit adds checking of the netlink error field and reports when
the deletion has failed.  When DeleteAll is called, we ignore the ENOENT
return since it could've raced somewhere and we don't actually care
since we're deleting all of the rules.