Add audit.ReceiveBuf to use caller-provided buffer
Receive uses an internal buffer to receive, process, and report the
message to the caller which means there's no way to decouple those
stages.  In event-heavy environments, this leads to messages getting
lost.  This commit introduces a new ReceiveBuf member that uses a
user-provided buffer. Receive now uses this with its internal buffer.